+++ ProjetoLecom/src/main/java/br/lecom/web/dao/service/IEstadoDAO.java
@@ -2,6 +2,6 @@ package br.lecom.web.dao.service;
 
 import br.lecom.web.vo.EstadoVO;
 
-public interface IEstadoDAO extends IEntityDAO<EstadoVO, Double>{
+public interface IEstadoDAO extends IEntityDAO<EstadoVO, Long>{
 
 }
