+++ ProjetoLecom/src/main/java/br/lecom/web/dao/service/IServicoEmpresaDAO.java
@@ -4,7 +4,7 @@ import java.util.List;
 
 import br.lecom.web.vo.ServicoEmpresaVO;
 
-public interface IServicoEmpresaDAO extends IEntityDAO<ServicoEmpresaVO, Double>{
+public interface IServicoEmpresaDAO extends IEntityDAO<ServicoEmpresaVO, Long>{
 	
 
 }
